finale yaklaşıyoruz cidden:)

+++ app/src/main/java/com/example/hadi_bakalm/model/SonIncelemeModel.java
@@ -1,13 +1,13 @@
 package com.example.hadi_bakalm.model;
 
 public class SonIncelemeModel {
-    private int id; // Room ID'si için eklendi
+    private long id; // int yerine long yapıldı (Room Entity uyumu için)
     private String baslik;
     private String aciklama;
     private String zaman;
-    private String tur; // "Kavram" veya "Metin"
+    private String tur;
 
-    public SonIncelemeModel(int id, String baslik, String aciklama, String zaman, String tur) {
+    public SonIncelemeModel(long id, String baslik, String aciklama, String zaman, String tur) {
         this.id = id;
         this.baslik = baslik;
         this.aciklama = aciklama;
@@ -15,7 +15,7 @@ public class SonIncelemeModel {
         this.tur = tur;
     }
 
-    public int getId() { return id; }
+    public long getId() { return id; }
     public String getBaslik() { return baslik; }
     public String getAciklama() { return aciklama; }
     public String getZaman() { return zaman; }
